requirements.txt
pdfplumber
python-dotenv
google-generativeai
streamlit